Use Godeps to manage external dependencies
Some of the dependencies like libovsdb and bitset come from our private forks
as we had made some fixes that haven't yet made it upstream in the respective
projects. We need to keep dependencies on these forks until the fixes are
merged upstream.

There are some forks (like that of etcd-client) that we haven't changed and
we can move our dependency to some working point in project's branch itself.
This shall be addressed as separate PR.

Signed-off-by: Madhav Puri <madhav.puri@gmail.com>