Go code checks can be run locally in container
Before the only option was to spin up a VM to run code checks, which is
expensive in terms of time.

Add Dockerfile-check to provide a container build for a runtime
environment that supports go tool vet, gofmt, and golint

Add script 'code_checks_in_docker.sh' intended to be run inside the
Dockerfile-check container that runs `make checks` which in turn runs a
variety of go code checks against packages passed in.

Checks are reordered so more severe issues (e.g. functional) are
resolved before less severe ones (e.g. formatting).

Some checks are upated to be simpler and faster because the command
accept multiple directories and recurse automatically.

Drive-by:
* removed some trailing whitespace
* added more exclusions in dockerignore
* variables for go check commands not needed, they are only refernced
  once and require a visual lookup to find what the command actually
  is

Signed-off-by: Chris Plock <chrisplo@cisco.com>