Contiv network is not deleted when docker network is deleted (#1029)
* Contiv network is not deleted when docker network is deleted

Steps to reproduce:
1. Create a docker network with docker network CLI
2. Ensure docker network and contiv network are created (docker network ls and netctl network ls)
3. Remove docker network.
4. Check docker network ls, docker network is deleted
5. Check netctl networ ls, contiv network is still test.

The root cause is due to the dockernetworkstate mapping is planned to delete after
contiv network is delete. However, the determination of contiv network deletion is
determined by the existance of the dockernetworkstate. Thefore, if the state is still
there, the contiv network will be unable to delete.

The fix is to delete the dockernetworkstate before proceed to contiv network deletion

Signed-off-by: Kahou Lei kalei@cisco.com

* Print out error

* Add unit test and also refactor some code to make the code more testable.

* Add more unit test and fix typos

* Add more testing

* Add key value pair in fake state

* Refactor base on comments

* Add more check in unit test.

Take out return statement when docker network is not there.