Init global settings from object store
Currently netplugin is trying to init from default settings if global
setting is not ready or wrong, and when the settings don't match it
will re-init netplugin. This is very easy to go wrong when there are
resources created before global settings are ready.
This commit makes netplugin wait until the global settings ready, and
initiate itself according to it. If the global settings are mis-match,
it will return error.
Also set default fwd mode to empty on netmaster and Refactor CIDRToMask

Signed-off-by: Wei Tie <wtie@cisco.com>